feat: moduleization
